requirements.txt
appdirs==1.4.3
asgiref==3.2.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
# Django==2.2.11
django-crispy-forms==1.9.0
django-heroku==0.3.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==2.1.0
social-auth-core==3.3.2
sqlparse==0.3.0
urllib3==1.25.8
virtualenv==20.0.10
whitenoise==5.0.1
